JS-5552: space create interface
