+++ src/scss/popup/common.scss
@@ -51,3 +51,4 @@ html.platformWindows {
 @import "./shortcut";
 @import "./usecase";
 @import "./share";
+@import "./spaceCreate";
