+++ src/scss/theme/dark/popup.scss
@@ -21,8 +21,7 @@
 			}
 		}
 
-		.side.right.tabSpaceIndex,
-		.side.right.tabSpaceCreate {
+		.side.right.tabSpaceIndex {
 			.sections {
 				.section {
 					.sectionContent {
