+++ src/ts/lib/util/space.ts
@@ -193,6 +193,32 @@ class UtilSpace {
 		return length < J.Constant.limit.space;
 	};
 
+	initSpaceState () {
+		const { widgets } = S.Block;
+		const blocks = S.Block.getChildren(widgets, widgets);
+
+		Storage.initPinnedTypes();
+
+
+		if (!blocks.length) {
+			return;
+		};
+
+		blocks.forEach(block => Storage.setToggle('widget', block.id, true));
+
+		const first = blocks[0];
+		const children = S.Block.getChildren(widgets, first.id);
+
+		if (children.length) {
+			const object = S.Detail.get(widgets, children[0].getTargetObjectId());
+
+			if (!object._empty_) {
+				Storage.setLastOpened(U.Common.getCurrentElectronWindowId(), { id: object.id, layout: object.layout });
+				U.Space.openDashboard('route');
+			};
+		};
+	};
+
 };
 
 export default new UtilSpace();
